requirements.txt
black==21.9b0
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.3.0
psycopg2==2.9.1
python-dotenv==0.19.0
regex==2021.8.28
requests==2.26.0
SQLAlchemy==1.4.25
tomli==1.2.1
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.1
